requirements.txt
asn1crypto==0.22.0
audioread==2.1.5
autopep8==1.3.3
bleach==2.0.0
cachetools==2.0.1
certifi==2017.7.27.1
cffi==1.10.0
chainer==2.1.0
chardet==3.0.4
conda==4.3.30
cryptography==2.0.3
cycler==0.10.0
decorator==4.1.2
dictdiffer==0.7.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.6
django-toolbelt==0.0.1
entrypoints==0.2.3
filelock==2.0.12
flake8==2.5.5
google-auth==1.1.1
gunicorn==19.7.1
hacking==1.0.0
html5lib==0.999999999
idna==2.6
ipaddress==1.0.18
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
jedi==0.10.2
Jinja2==2.9.6
joblib==0.11
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
kubernetes==3.0.0
librosa==0.5.1
llvmlite==0.20.0
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.2.1
mistune==0.7.4
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.0.0
numba==0.35.0
numpy==1.13.3
olefile==0.44
openshift==0.3.3
pandas==0.20.3
pandocfilters==1.4.2
pbr==3.1.1
pep8==1.5.7
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.4.0
psycopg2==2.7.3.2
ptyprocess==0.5.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycodestyle==2.3.1
pycosat==0.6.2
pycparser==2.18
pyflakes==0.8.1
Pygments==2.2.0
pyOpenSSL==17.2.0
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.6.1
python-string-utils==0.6.0
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.18.4
resampy==0.2.0
rsa==3.4.2
ruamel-yaml==0.11.14
ruamel.yaml==0.15.34
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.8.1
simplegeneric==0.8.1
six==1.10.0
static3==0.7.0
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.40.0
widgetsnbextension==3.0.2
xgboost==0.6a2
